+++ app/utils/crypto/symmetric.rb
@@ -3,7 +3,16 @@
 require 'openssl'
 require 'digest/sha1'
 
-class Crypto::Symmetric
+class ::Crypto::Symmetric
+  class_attribute :password_bitsize
+
+  LATEST_ALGORITHM = 'AES256IV'
+
+  # Add further algorithms at the bottom
+  ALGORITHMS = {
+    AES256: ::Crypto::Symmetric::Aes256,
+    AES256IV: ::Crypto::Symmetric::Aes256iv
+  }.with_indifferent_access.freeze
 
   class << self
 
@@ -17,6 +26,14 @@ class Crypto::Symmetric
 
     def random_key
       raise 'Implement in subclass'
+    end
+
+    def all_algorithms
+      ALGORITHMS.keys
+    end
+
+    def latest_algorithm?(entry)
+      LATEST_ALGORITHM == entry.encryption_algorithm
     end
   end
 end